(PC-23710)[PRO] refactor: move PeriodSelector out of form_raw
