+++ pro/src/screens/Offers/SearchFilters/SearchFilters.tsx
@@ -16,9 +16,9 @@ import fullRefreshIcon from 'icons/full-refresh.svg'
 import strokeCloseIcon from 'icons/stroke-close.svg'
 import { ButtonLink } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
+import PeriodSelector from 'ui-kit/form/PeriodSelector/PeriodSelector'
 import SelectInput from 'ui-kit/form/Select/SelectInput'
 import { BaseInput, FieldLayout } from 'ui-kit/form/shared'
-import PeriodSelector from 'ui-kit/form_raw/PeriodSelector/PeriodSelector'
 import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
 
 import styles from './SearchFilters.module.scss'
