+++ pro/src/ui-kit/form/PeriodSelector/PeriodSelector.module.scss
