+++ pro/src/ui-kit/form/PeriodSelector/__specs__/PeriodSelector.spec.tsx
